Split laserscan and robot pose handling in two nodes
