+++ build.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-mkdir -p listener/build
-pushd listener/build
+mkdir -p nodes/build
+pushd nodes/build
 cmake ..
 make
 popd
