+++ run.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
 
-gazebo worlds/eurobot2016.world
-./listener/build/listener
+gazebo worlds/eurobot2016.world && ./nodes/build/laserscan && ./nodes/build/robotpose
